% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{onnx_is_installed}
\alias{onnx_is_installed}
\title{Check whether ONNX Runtime is available}
\usage{
onnx_is_installed()
}
\value{
\code{TRUE} if the ONNX Runtime shared library can be found
in any of the standard search locations, \code{FALSE} otherwise.
}
\description{
Check whether ONNX Runtime is available
}
\examples{
onnx_is_installed()
}
