% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{ort_is_installed}
\alias{ort_is_installed}
\title{Check whether ONNX Runtime is available}
\usage{
ort_is_installed()
}
\value{
`TRUE` if the ONNX Runtime shared library can be found
  in any of the standard search locations, `FALSE` otherwise.
}
\description{
Check whether ONNX Runtime is available
}
\examples{
ort_is_installed()
}
