% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{ort_is_loaded}
\alias{ort_is_loaded}
\title{Check whether ONNX Runtime is loaded}
\value{
`TRUE` if the ONNX Runtime shared library has been loaded
  in the current R session, `FALSE` otherwise.
}
\description{
Check whether ONNX Runtime is loaded
}
\examples{
ort_is_loaded()
}
